+++ src/main.rs
@@ -1,3 +1,5 @@
+#[macro_use] extern crate lazy_static;
+
 use std::cmp;
 use clap::{Arg, App, AppSettings, SubCommand};
 // TEMPORARY_VENDOR
@@ -9,15 +11,21 @@ use colored::*;
 use chrono::{DateTime, Local, Utc};
 use num_format::{Locale, ToFormattedString};
 
-const VERSION: &'static str = env!("CARGO_PKG_VERSION");
-const APPNAME: &'static str = env!("CARGO_PKG_NAME");
 
+const VERSION: &'static str = env!("CARGO_PKG_VERSION");
+const ORGANIZATION: &'static str = "joseluis";
+const APPNAME: &'static str = env!("CARGO_PKG_NAME");
+const CONFIGNAME: &'static str = "config.toml";
+
+mod configuration;
+use configuration::*;
 
 fn main() {
     let client = cia::SyncClient::new();
 
-    // DEFINE THE ARGUMENTS
+    Settings::read();
 
+    // DEFINE THE ARGUMENTS
     let argmatch = App::new(APPNAME)
         .version(VERSION)
         .setting(AppSettings::VersionlessSubcommands)
@@ -96,6 +104,152 @@ fn main() {
             )
             */
         )
+
+
+        .subcommand(SubCommand::with_name("list")
+            .about("Manage your lists of crates")
+
+            .subcommand(SubCommand::with_name("show")
+                .help("shows the crates contained in the list")
+                .arg(Arg::with_name("list")
+                     .required(false)
+                     .empty_values(false)
+                )
+                .arg(Arg::with_name("info")
+                     .short("i")
+                     .long("info")
+                     .help("show the information of each crate in the list")
+                     .required(false)
+                     .multiple(true)
+                     .empty_values(true)
+                 )
+            )
+            .subcommand(SubCommand::with_name("add")
+                .help("add a crate to a list")
+                .arg(Arg::with_name("list")
+                     .help("the list where to add the crate")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("crate")
+                     .help("the crate to add to the list")
+                     .required(true)
+                     .empty_values(false)
+                     //.multiple(true) // TODO: allow multiple
+                     .index(2)
+                )
+            )
+            .subcommand(SubCommand::with_name("new")
+                .help("create a new empty list")
+                .arg(Arg::with_name("list")
+                     .required(true)
+                     .empty_values(false)
+                     //.multiple(true) // TODO: allow multiple
+                )
+            )
+            .subcommand(SubCommand::with_name("del")
+                .help("delete an empty list")
+                .arg(Arg::with_name("list")
+                     .help("The list to delete (must be empty)")
+                     .required(true)
+                     .empty_values(false)
+                     //.multiple(true) // TODO: allow multiple
+                )
+            )
+            .subcommand(SubCommand::with_name("rem")
+                .help("remove a crate from a list")
+                .arg(Arg::with_name("list")
+                     .help("The list containing the crate")
+                     .required(true)
+                     .empty_values(false)
+                )
+                .arg(Arg::with_name("crate")
+                     .help("the crate to remove")
+                     .required(false)
+                     .empty_values(false)
+                     //.multiple(true) // TODO: allow multiple
+                )
+            )
+
+            /*
+            .subcommand(SubCommand::with_name("copy")
+                .help("copy a crate from one list to another")
+                .arg(Arg::with_name("list_from")
+                     .help("the list and crate to copy from (list:crate)")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("list_to")
+                     .help("the list where to copy the crate")
+                     .required(true)
+                     .empty_values(false)
+                     .index(2)
+                )
+            )
+            .subcommand(SubCommand::with_name("move")
+                .help("move a crate from one list to another")
+                .arg(Arg::with_name("list_from")
+                     .help("the list and crate to move from (list:crate)")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("list_to")
+                     .help("the list where to move the crate")
+                     .required(true)
+                     .empty_values(false)
+                     .index(2)
+                )
+            )
+            .subcommand(SubCommand::with_name("copy-all")
+                .help("copy all crates from one list to another")
+                .arg(Arg::with_name("list_from")
+                     .help("the list from where all the crates will be copied")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("list_to")
+                     .help("the list where to copy all the crates")
+                     .required(true)
+                     .empty_values(false)
+                     .index(2)
+                )
+            )
+            .subcommand(SubCommand::with_name("move-all")
+                .help("move all crates from one list to another")
+                .arg(Arg::with_name("list_from")
+                     .help("the list from where all the crates will be moved")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("list_to")
+                     .help("the list where to move all the crates")
+                     .required(true)
+                     .empty_values(false)
+                     .index(2)
+                )
+            )
+            .subcommand(SubCommand::with_name("clone")
+                .arg(Arg::with_name("list_existing")
+                     .help("the list to clone")
+                     .required(true)
+                     .empty_values(false)
+                     .index(1)
+                )
+                .arg(Arg::with_name("list_new")
+                     .help("the new cloned list (must not exist)")
+                     .required(true)
+                     .empty_values(false)
+                     .index(2)
+                )
+            )
+            */
+            //
+        )
         .get_matches();
 
 
@@ -140,6 +294,113 @@ fn main() {
                     _ => unreachable!()
                 }
             },
+
+        // LIST ARGUMENTS
+
+        ("list", Some(list_matches)) => {
+            match &list_matches.subcommand() {
+                ("show", Some(args)) => {
+                    if let Some(list) = args.value_of("list") {
+                        if Lists::exists(list) {
+                            println!("Your list \"{}\" contains {} crates:",
+                                list.bright_green(), Lists::quantity(list));
+
+                            match args.occurrences_of("info") {
+                                0 => if let Some(contents) = Lists::show(list, false) {
+                                    println!("{}", contents);
+                                }
+                                // TODO: move this to 2 or more occurences, and
+                                // make a more compact presentation for 1 occurrence
+                                1 | _ => if let Some(contents) = Lists::show(list, true) {
+                                    for crate_name in contents.split_whitespace() {
+                                        let _ = show_crate(&client, crate_name,
+                                            args.occurrences_of("reverse"));
+                                        println!("");
+                                    }
+                                }
+                            }
+
+                        } else {
+                            println!("List \"{0}\" doesn't exist. You can create it with '{1}'",
+                                list.bright_red(), format!("crin list new {}",
+                                    list.bright_green()).bright_blue());
+                        }
+
+                    } else {
+                        // if no list is provided, show which lists there are
+                        match args.occurrences_of("info") {
+                            0 => Lists::show_lists(false),
+                            1 | _ => Lists::show_lists(true),
+                        }
+                    }
+                },
+                ("new", Some(args)) => {
+                    if let Some(list) = args.value_of("list") {
+                        println!("creating: {}", list);
+                        Lists::new(list);
+                    }
+                },
+                ("del", Some(args)) => {
+                    // TODO: allow multiple
+                    if let Some(list) = args.value_of("list") { Lists::del(list); }
+                },
+                ("add", Some(args)) => {
+                    // TODO: allow multiple
+                    Lists::add(args.value_of("list").unwrap(), args.value_of("crate").unwrap());
+                },
+                ("rem", Some(args)) => {
+                    // TODO: allow multiple
+                    Lists::rem(args.value_of("list").unwrap(), args.value_of("crate").unwrap());
+                },
+                /*
+                // TODO:
+                ("copy", Some(args)) => {
+                    let list_from = args.value_of("list_from").unwrap();
+                    let list_to = args.value_of("list_to").unwrap();
+
+                    if list_from == list_to {
+                        println!("You must copy between different lists");
+                    } else {
+                        // TODO check both lists exist, and crate too
+                        println!("copy: {} to {}", list_from, list_to);
+                    }
+                },
+                // TODO:
+                ("move", Some(args)) => {
+                    let list_from = args.value_of("list_from").unwrap();
+                    let list_to = args.value_of("list_to").unwrap();
+
+                    if list_from == list_to {
+                        println!("You must move between different lists");
+                    } else {
+                        println!("move: {} to {}", list_from, list_to);
+                    }
+                }
+                // TODO:
+                ("move-all", Some(args)) => {
+                    let list_from = args.value_of("list_from").unwrap();
+                    let list_to = args.value_of("list_to").unwrap();
+
+                    if list_from == list_to {
+                        println!("You must move between different lists");
+                    } else {
+                        println!("move all crates from: {} to {}", list_from, list_to);
+                    }
+                }
+                // TODO:
+                ("clone", Some(args)) => {
+                    let list_existing = args.value_of("list_existing").unwrap();
+                    let list_new = args.value_of("list_new").unwrap();
+
+                    // TODO check the list exist and the other doesn't
+                    println!("cloning from {} to {}", list_existing, list_new);
+                },
+                */
+                ("", None) => { Lists::show_lists(false); }
+                _ => unreachable!(),
+            }
+        }
+
         _ => println!("{} help", APPNAME),
     }
 }
@@ -574,4 +835,19 @@ fn date_ago(date: &DateTime<Utc>, num_items: usize) -> String {
         .to_string())
 }
 
+/// Return a string of words separated by commas,
+/// optionally surrounding each word with a string.
+fn commify(words: Vec<&str>, surround: &str,
+    _color_word: Option<Color>, _color_comma: Option<Color>) -> String {
+    let mut text = "".to_string();
+    for word in words {
+        text = format!("{t}{c} {s}{w}{s}", t = text,
+            s = surround,
+            // TODO: use the optional colors for words and commas
+            c = ",",
+            w = word
+        );
+    }
+    return text[1..].trim().to_string(); // remove the leading comma
+}